requirements.txt
loguru
pyTelegramBotAPI
requests
telebot-calendar
python-dotenv
peewee
